Add FirebaseAuthenticationBackend for graphql auth
Add a new authentication backend class: FirebaseAuthenticationBackend for doing
authentication using firebase sdk auth. We expect to have an HTTP_AUTHORIZATION
header in the request with a token from firebase. We validate than token to get
the user from our database and then return the user otherwise a None is returned.